+++ src/Samples/DIPS.Xamarin.UI.Samples/Controls/ContextMenu/ContextMenuPage.xaml
@@ -9,15 +9,19 @@
         <menu1:ContextMenuPageViewModel />
     </ContentPage.BindingContext>
     <StackLayout Margin="15">
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+        <!-- Single item -->
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
                 <contextMenu:ContextMenuItem Title="Item" />
             </contextMenu:ContextMenuControl.ItemsSource>
             <Button Text="Single item"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
-        
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+
+        <!-- Two items -->
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
                 <contextMenu:ContextMenuItem Title="Item 1" />
                 <contextMenu:ContextMenuItem Title="Item 2" />
@@ -25,17 +29,23 @@
             <Button Text="Two items"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
-        
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+
+        <!--Checkable items (all)-->
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
-                <contextMenu:ContextMenuItem Title="Item 1" IsCheckable="True"/>
-                <contextMenu:ContextMenuItem Title="Test 2" IsCheckable="True"/>
+                <contextMenu:ContextMenuItem Title="Item 1"
+                                             IsCheckable="True" />
+                <contextMenu:ContextMenuItem Title="Test 2"
+                                             IsCheckable="True" />
             </contextMenu:ContextMenuControl.ItemsSource>
             <Button Text="Checkable items (all)"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
-        
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+
+        <!--Checkable items (single)-->
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
                 <contextMenu:ContextMenuGroup IsCheckable="True">
                     <contextMenu:ContextMenuItem Title="Item 1" />
@@ -45,37 +55,41 @@
             <Button Text="Checkable items (single)"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
-        
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
                 <contextMenu:ContextMenuGroup Title="Group 1">
-                    <contextMenu:ContextMenuItem Title="Item 1"/>
-                    <contextMenu:ContextMenuItem Title="Item 2"/>
+                    <contextMenu:ContextMenuItem Title="Item 1" />
+                    <contextMenu:ContextMenuItem Title="Item 2" />
                 </contextMenu:ContextMenuGroup>
                 <contextMenu:ContextMenuGroup Title="Group 2">
-                    <contextMenu:ContextMenuItem Title="Item 1"/>
-                    <contextMenu:ContextMenuItem Title="Item 2"/>
+                    <contextMenu:ContextMenuItem Title="Item 1" />
+                    <contextMenu:ContextMenuItem Title="Item 2" />
                 </contextMenu:ContextMenuGroup>
             </contextMenu:ContextMenuControl.ItemsSource>
             <Button Text="Grouped items"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
-        
-        <contextMenu:ContextMenuControl VerticalOptions="Start">
+
+        <contextMenu:ContextMenuControl VerticalOptions="Start"
+                                        ItemClickedCommand="{Binding ItemClickedCommand}">
             <contextMenu:ContextMenuControl.ItemsSource>
-                <contextMenu:ContextMenuGroup Title="Group 1" IsCheckable="True">
-                    <contextMenu:ContextMenuItem Title="Item 1"/>
-                    <contextMenu:ContextMenuItem Title="Item 2"/>
+                <contextMenu:ContextMenuGroup Title="Group 1"
+                                              IsCheckable="True">
+                    <contextMenu:ContextMenuItem Title="Item 1" />
+                    <contextMenu:ContextMenuItem Title="Item 2" />
                 </contextMenu:ContextMenuGroup>
-                <contextMenu:ContextMenuGroup Title="Group 2" IsCheckable="True">
-                    <contextMenu:ContextMenuItem Title="Item 1"/>
-                    <contextMenu:ContextMenuItem Title="Item 2"/>
+                <contextMenu:ContextMenuGroup Title="Group 2"
+                                              IsCheckable="True">
+                    <contextMenu:ContextMenuItem Title="Item 1" />
+                    <contextMenu:ContextMenuItem Title="Item 2" />
                 </contextMenu:ContextMenuGroup>
             </contextMenu:ContextMenuControl.ItemsSource>
             <Button Text="Checkable grouped items"
                     BackgroundColor="White" />
         </contextMenu:ContextMenuControl>
     </StackLayout>
-    
-    
+
+
 </ContentPage>
