+++ src/Samples/DIPS.Xamarin.UI.Samples/Controls/ContextMenu/ContextMenuPage.xaml.cs
@@ -11,14 +11,5 @@ namespace DIPS.Xamarin.UI.Samples.Controls.ContextMenu
         {
             InitializeComponent();
         }
-
-        protected override void OnAppearing()
-        {
-            base.OnAppearing();
-            if (BindingContext is ContextMenuPageViewModel contextMenuPageViewModel)
-            {
-                _ = contextMenuPageViewModel.Initialize();
-            }
-        }
     }
 }
